refactor(cli): extract unity qa handler
